requirements-production.txt
# ALC-Algo Production Requirements
# Install with: pip install -r requirements-production.txt

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Data & Analytics
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0

# Azure
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-ai-ml==1.12.0
applicationinsights==0.11.10

# Market Data
polygon-api-client==1.12.0
alpaca-py==0.12.0

# AI/ML APIs
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
pydantic==2.5.3

# Monitoring & Dashboards
streamlit==1.29.0
plotly==5.18.0

# Scheduling
schedule==1.2.0
